backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.2
Pillow==10.1.0
opencv-python-headless==4.8.1.78  # Use headless version
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
huggingface-hub==0.19.4
datasets==2.14.6
accelerate==0.24.1
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
pytesseract==0.3.10
pdf2image==1.16.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1